pyproject.toml
[tool.poetry]
name = "iso-standards-games"
version = "0.1.0"
description = "Interactive educational games for learning ISO standards"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "iso_standards_games"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
httpx = "^0.26.0"
python-i18n = "^0.3.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
select = ["E", "F", "I", "UP"]
line-length = 88
target-version = "py39"